backend/requirements.txt
Flask
flask-restx
psycopg2-binary
peewee
pandas
python-dotenv
flask-cors